refactor: move netstring to internal
