+++ packages/internal/src/netstring.js
